pyproject.toml
[build-system]
requires = ["setuptools>=61.2"]
build-backend = "setuptools.build_meta"

[project]
name = "hwdesk"
version = "1.0.3"
dependencies = [
    "ch9329==1.2.0",
    "keyboard==0.13.5",
    "opencv-cam-idx==1.0.1",
    "opencv-python==4.9.0.80",
    "pillow==10.3.0",
    "pydantic==2.6.4",
]
requires-python = ">=3.11"
authors = [{ name = "Pradish Bijukchhe", email = "pradish@sandbox.com.np" }]
description = ""
readme = "README.md"
license = { file = "LICENSE" }
keywords = []
classifiers = ["Programming Language :: Python :: 3"]

[project.urls]
Homepage = "https://github.com/sandbox-pokhara/hwdesk"
Issues = "https://github.com/sandbox-pokhara/hwdesk/issues"

[project.gui-scripts]
hwdesk = "hwdesk.__main__:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-dir]
"hwdesk" = "hwdesk"

[tool.isort]
line_length = 79
force_single_line = true

[tool.black]
line-length = 79
preview = true

[tool.pyright]
typeCheckingMode = "strict"
include = ["hwdesk", "test.py"]
